requirements.txt
fastapi
uvicorn
sentence-transformers
chromadb
mysql-connector-python
python-multipart
numpy